# .specify/specs/wechat-login-credit-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: WeChat Login and Credit System API
  description: API for user authentication and credit management system
  version: 1.0.0
  contact:
    name: AI Copywriting Assistant
    url: https://github.com/your-org/copywriting-assistant
servers:
  - url: https://api.copywriting-assistant.com
    description: Production server
  - url: https://staging-api.copywriting-assistant.com
    description: Staging server
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
  /api/auth/wechat/qr:
    post:
      summary: Generate WeChat QR code for login
      description: Creates a QR code for WeChat OAuth login
      tags:
        - Authentication
      security: []
      responses:
        '200':
          description: QR code generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  qr_url:
                    type: string
                    format: uri
                    description: QR code image URL
                  state:
                    type: string
                    description: OAuth state parameter
                  expires_at:
                    type: string
                    format: date-time
                    description: QR code expiration time
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/wechat/callback:
    post:
      summary: Handle WeChat OAuth callback
      description: Processes WeChat OAuth callback and creates user session
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  description: Authorization code from WeChat
                state:
                  type: string
                  description: OAuth state parameter
              required:
                - code
                - state
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  session:
                    $ref: '#/components/schemas/Session'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/wechat/status:
    get:
      summary: Check WeChat login status
      description: Polls for WeChat login completion
      tags:
        - Authentication
      security: []
      parameters:
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: OAuth state parameter
      responses:
        '200':
          description: Login status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [pending, completed, expired, failed]
                  user:
                    $ref: '#/components/schemas/User'
                  session:
                    $ref: '#/components/schemas/Session'
        '400':
          description: Invalid state parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        wechat_openid:
          type: string
        wechat_unionid:
          type: string
          nullable: true
        nickname:
          type: string
        avatar_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Session:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: supabase-auth-token
...